pyproject.toml
[project]
name = "mec202"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "opencv-python>=4.13.0.92",
    "paddleocr<3",
    "paddlepaddle<3",
    "pillow>=12.2.0",
    "pyserial>=3.5",
    "pyzbar>=0.1.9",
    "requests>=2.33.1",
    "scikit-image>=0.26.0",
    "werkzeug>=3.1.8",
]
